+++ RClick.xcodeproj/project.xcworkspace/xcuserdata/lixu.xcuserdatad/WorkspaceSettings.xcsettings
@@ -3,11 +3,17 @@
 <plist version="1.0">
 <dict>
 	<key>BuildLocationStyle</key>
-	<string>UseAppPreferences</string>
+	<string>CustomLocation</string>
+	<key>CustomBuildIntermediatesPath</key>
+	<string>Build/Intermediates.noindex</string>
 	<key>CustomBuildLocationType</key>
 	<string>RelativeToDerivedData</string>
+	<key>CustomBuildProductsPath</key>
+	<string>Build/Products</string>
+	<key>DerivedDataCustomLocation</key>
+	<string>DerivedData</string>
 	<key>DerivedDataLocationStyle</key>
-	<string>Default</string>
+	<string>WorkspaceRelativePath</string>
 	<key>ShowSharedSchemesAutomaticallyEnabled</key>
 	<true/>
 </dict>
