+++ RClick/Localizable.xcstrings
@@ -201,16 +201,6 @@
         }
       }
     },
-    "Open With %@" : {
-      "localizations" : {
-        "zh-Hans" : {
-          "stringUnit" : {
-            "state" : "translated",
-            "value" : "用 %@ 打开"
-          }
-        }
-      }
-    },
     "Quit" : {
       "localizations" : {
         "en" : {
